[FIX] l10n_mx_edi_extended: fix MemoryError on install
When there are too many account move lines installing this module may
cause a MemoryError. In order to avoid the issue with pre-compute the
values that otherwise will be compted by the ORM.

For reference, the change from previous stored value[^1] (not computed)
to computed stored values[^2] was done on #12226.

opw-2796094

[^1]: https://github.com/odoo/enterprise/commit/1ccb480b8e95f5702b6fb4e01446d24107514550#diff-8afc82e663ecf1ebc79d464ba7d9e96a7fda9fe2b431abecbd4f7216114de2dfL39-L45
[^2]: https://github.com/odoo/enterprise/commit/1ccb480b8e95f5702b6fb4e01446d24107514550#diff-134267cacedaacc196d4520947da5a6271c551ddb3d7b0a9867ea6b46402bdb4R94-R105

closes odoo/enterprise#25959

X-original-commit: a73a8c1fce8077b859e6a088ab1252b86265a63a
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>
Signed-off-by: Alvaro Fuentes <afu@odoo.com>