[FIX] web_gantt: prevent indeterministic test fail
Since fc5179a3 `ganttViewParams` which is used for all the tests
is modified during the test on `collapse_first_level`. As such, it
is potentially breaking tests that are performed after as they were
relying on the fact that `collapse_first_level` is not set.

This commit prevent the mutation of `ganttViewParams`.

This commit also:
- Adds missing `;` in tests.
- Modifies some of the assert in order to improve debugging when
  test fails.

closes odoo/enterprise#25733

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>