[FIX] industry_fsm: set the default FSM project from map menu
Before this commit, when the user creates a fsm task from
`field service > map menu`, a fsm project with the lowest sequence
should be selected and set by default on that new task.
The current behavior is no project is set to the task and so the
task is considered as a private task by default.

This commit adds `fsm_mode=True` in the context of the action to
search the fsm project with the lowest sequence instead of creating
a private task by default.

task-2788878

closes odoo/enterprise#28338

X-original-commit: c2c78a427ccbd20523d1bc94ffba69b1eb78e274
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>