[FIX] stock_barcode: scan package level
Before this commit, when a transfert operation is using the "Move Entire
Package" setting and the package level is scanned in the Barcode App,
the package line wasn't correctly updated.

How to reproduce:
  - Enable package (inventory settings);
  - For the delivery operation type, mark the "Move Entire Package"
    setting;
  - Create a product and add some quantity on hand inside a package;
  - Create a delivery and in its package level line, add the create
    package;
  - Confirm the delivery and open it in the Barcode App;
    -> You should have only one line for the package (0/1, with a button
    to see the package's content.)
  - Scan the package -> The package line isn't update (still 0/1).

Actually, when the package is scanned, its quants were retrieve and move
lines are created for it. But what we really want is the package level's
move line to be updated, not new ones created.

X-original-commit: 4eb77be4e40dfb77aae009322d6a93d01b9159df
Part-of: odoo/enterprise#25941