[FIX] appointment: fix custom appointment duration days
The difference between the reference_date and the last unique slot
does not represent always the correct duration because of the datetime.
Based on the hours, the delta between the two datetime could lead to a
duration shorten by a day.

For example, if the reference date and the last unique slot are less than
1 day apart, the appointment_duration_days was equal to 0.
(reference_date = 1st January 14:00 and last_slot = 2nd January 09:00)
This was leading to weird behaviour, where a slot already taken
was still displayed in the availabilities.

Indeed the appointment_duration_days would than be equal to 0 because
the 'days' property of the time delta would be equal to 0.
This didn't reflect the correct duration and was impacting the
availabilities computation.

task-2806736

closes odoo/enterprise#26746

X-original-commit: 9a27afb14c0da2af926b9d988d1dbd4f85cd9ccb
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>