[IMP] *appointment*: rename route
Rename the routes related to the appointment module and begining
with "/calendar" into "/appointment". The goal is to better match
the naming of the module and to prepare the changes of the public
access_rights on the appointment type model.

We keep for backward compatibility the old routes:
"/calendar" and "/calendar/<appointment_type_id>" in order to have
all previous links working as intended.

We also remove a route deprecated since quite a long time:
"/calendar/<model('appointment_type'):appointment_type/appointment"
(see odoo/enterprise#9456)

taks-2427015

Part-of: odoo/enterprise#23646