[IMP] *appointment*: share unpublished appointment with invite
when an appointment type is unpublished, it can now be shared
thanks to an invitation link.
This invitation is linked to a token that allows to store the
different params of the appointment link.

This allows to share easily single and multiple appointment types
and to have a way to known from which invitation an appointment
has been booked to manage correctly the cancellation.

Now public user cannot read all appointment types without the website
module. They are no more considered published in the appointment
module.

We also display the unpublished appointment types in the "Share
Availabilities" dropdown, which create now an invitation.

task-2716094

Part-of: odoo/enterprise#23646