[IMP] appointment(_hr): quickly cleanup some tests
Purpose is to cleanup some leftover not already cleaned in odoo/enterprise#24900
A duplicated test (``test_timezone_delta``) is removed.

We also add a test for a corner case to fix: for "today" we should not
generate slots if it is actually before the reference date. Future commit
will solve that issue and allow the test to be green.

Task-2695730

Part-of: odoo/enterprise#23298