[FIX] mrp_plm: changed filter domain on a computed filed
Reproduction:
1. Go to PLM-> configuration -> ECO stages, create 2 approvals for stage
New, one for Mitchell Admin and the other for Marc Demo
2. Create an ECO and approve it as Michell Admin
3. Check the filter “Awaiting My Validation” in ECO kanban view, the ECO
is still there

Reason: The filter domain wasn’t correct. Each ECO has a many2many field
approval_ids. The issue is caused because we check all the approvals
from all stages. But we only need to check the approvals which are at
the same stage as the ECO, e.g. approval.template_stage_id ==
eco.stage_id. To check if the user_id is in their many2many field
required_user_ids, and this approval is not closed yet (still needs to
be approved).

Fix: Added a new compute field awaiting_my_validation for this filter in
MrpEcoApproval class, and search domain for this filter is simply
awaiting_my_validation = True. Its computation method
_compute_awaiting_my_validation will trigger the search method
_search_awaiting_my_validation, then the value awaiting_my_validation is
set to True to those approvals satisfying the search criteria. For other
approvals, awaiting_my_validation is False. In searching method
_search_awaiting_my_validation, it returns a domain where the approval
is not closed, not approved, having the current user as
required_user_ids, and with approval_template_id.approval_type is
mandatory or optional

opw-2802238

closes odoo/enterprise#27009

X-original-commit: 7fd12bc64f485f20ededfde8b98e4bd6c039f7c0
Signed-off-by: Arnold Moyaux <arm@odoo.com>