[IMP] sale_timesheet_enterprise: add grid view for timesheets action
Before this commit, the grid view is not shown in the action used for
Sales section shown in the right side panel in project update kanban
view.

This commit adds this grid in that action and change the form view to
use the form view defined in `timesheet_grid` module instead of the
one in `hr_timesheet` one.

task-2710808