[FIX] helpdesk: allow correct date filtering on analysis reports
In both SLA Status Analysis and Ticket Analysis reports, create date is
set as a Date field only, while it is in fact a Datetime field inherited
from the helpdesk.ticket object in the SQL view.

This has the consequence that we cannot filter properly the information
using dates : if we set the filter to Creation Date = current month, all
tickets created on the last day of this month won't be retrieved, hence
providing incorrect results.

closes odoo/enterprise#26554

X-original-commit: 1097fc59e303a9b2f6be9d875fe3599100865d88
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>