[FIX] account_bank_statement_import_camt: no TtlAmt element
Since commit [1], we added a check between the amount retrieved from the
details of an operation and the total amount found in the batch
information in _total_amount_getters. However, if no TtlAmt is found, we
wrongly compare the operation amount with the entry amount. This cannot
work if there is more than one operation in the batch. In case TtlAmt
element is not present, we should just return the retrieved amount without
further additional checks.

[1]: https://github.com/odoo/enterprise/commit/df63265fe2b310ca20b92097383228e78ee1da58

closes odoo/enterprise#25966

X-original-commit: 29b00fad23a07504e14805c3ab4a9d997bf03386
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>