[IMP] appointment: Allow to define slot until midnight
Previously, midnight couldn't be set as end of a slot, 23:59 was used instead.
With this change, 00:00 can be used as 00:00 of the next day.

This change allows users to effectively generate slots "until midnight", which
was not possible before.

Technical note:

Method AppointmentSlot._convert_end_hour_24_format has been added to facilitate
checks and computations.

We haven't converted it as a computed field because it is merely used in
compute and checks which would have required to add not straightforward
dependencies to those methods.

Task-2810493

closes odoo/enterprise#25827

Related: odoo/upgrade#3470
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>