[FIX] hr_holidays_gantt: fix leaves' planning synchronisation
Observed Behaviour

When adding a time off on a period that include a public holiday, the time off
day will end on the public holiday according to the Planning of viewing the
Work Entries.

Expected Behaviour

When creating a public holiday on a day and approving time off for a period
that include this public holiday, the planning should simply include it in
the whole period of the holiday.

Reproducibility

This issue can be reproduced using the following steps:
1. Add a public holiday on a monday (eg May 23rd)
2. Request Time off from the previous friday to the next tuesday (ie May 20-24)
3. Validate the Time Off
4. Go to the Planning and plan a new task for the month including the time off
It will show that the time off is ending on the public holiday (ie May 20-23)

Fix Description

When looping on the holiday to determine the availbility of the user, we check
if a holiday's start is included in/next to another holiday and, if so, we
set the full period as the begginning of the first holiday to the end of the
second, regardless if the end of the first one is before/after the end of the
second one. We add a condition to check this before changing the end date.

Relate issues/PR
opw-2797548

closes odoo/enterprise#30642

X-original-commit: 3e54b2237866253c24970ea7fa48191a53f6fd21
Signed-off-by: Kevin Baptiste <kba@odoo.com>