[FIX] worksheet: worksheet template deletion error
Step to reproduce:
-Install field service
-Activate worksheets in configuration settings
-Try to delete a worksheet template in configuration> worksheet templates

Expected behavior:
The template is deleted

Current behavior:
There is an saying the template does not exists

Explanation:
The worksheet template was deleted twice, and the second deletion generates
the error. The first deletion was due to the deletion of the model linked to
the worksheet template via the model_id field, this field was "ondelete=cascade"
and hence the deletion of the model also deleted the worksheet template.
The secund deletion was the classical call super.unlink(). To prevent the
error we add a self.exists() in the super call that will filter deleted records.

opw-2810142

closes odoo/enterprise#26067

X-original-commit: c17fb7ebe5001dbf370db2c5345e528bfcd4f2cf
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>