[FIX] planning: set the company working date as default date
before this commit, the planning.slot gantt view > add button > the default date
was showing the whole week as working days of the company when we switch
to the previous week in planning gantt view.

so after this commit, the default date is set based on the
company's working calendar.

task-2790017

closes odoo/enterprise#28124

X-original-commit: b66fcb144b1c5efebf8a513879fa323129d1dbb5
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>