[FIX] appointment: apply assign method on filtered appointment time selection
1) ASSIGN METHOD COHERENCE

Before, when picking a time on an appointment_type with 'random' assign method,
with filter_staff_user_ids value set, the user was able to select in the dropdown
the user they wanted among the suggested ones. This is not coherent and we should
not have to choose between the filtering and the assign method.

2) RESTRICT STAFF USERS TO THE SUGGESTED VALUES (existing values based on filter):

Fixing point 1) reveals an issue: if the appointment_type is randomly assigned,
then the user select dropdown is now absent. A single user is not linked to the view,
but potentially the whole filtered selection.

However, the method updating the slots availabilities is called when either the TZ
or the assigned user (in the dropdown) changes. As the staff_user is not set if
the dropdown is not in the view, the method is called with staff_user=False, leading
to using staff_user = False in the rest of the computation.

Therefore, if we change the tz - still in the view - the slots will be updated
using all staff_users on the appointment (default behaviour, when staff_user=False),
erasing in practice the filter on staff users. For instance I have received a link
with filter_staff_user_ids = [2,3], while users [2,3,4] are set on the appointment
type. 2 and 3 are not available next week. I change timezone and slots appear next
week, because user 4 is available. We want to avoid that.

-> SOLUTION

We fix those linked issues by

- Giving the main methods that compute the availabilities
an 'extension': staff_user argument can now contain multiple users, acting as a
restriction - if set - to only consider those users. This will contain the possible
users (after filtering). It will still be the chosen user in the dropdown if assign
method is chosen. Therefore, the methods can take into account the filtered users.
This argument will be renamed in master to filter_users to represent that
there can be more than one staff user and that it is used to filter.

- Remove the staff user select dropdown from the view if the assign method is not
'chosen'. Therefore the staff_user is not set when calling the update function
(e.g. when changing tz), but the update function will now filter the staff users
if 'filter_staff_user_ids' is given to tackle issue 2) (again, by default, calling
the slots method with staff_user = False uses all users of the appointment)

TESTS

A test is added to ensure that behaviour.

Task-2496629

Part-of: odoo/enterprise#24744