[FIX] stock_barcode: pages' order
When multilocation is active, a barcode operation can have multiple
pages: one for each location.
Before this commit, these pages were sorted by their location's id which
isn't the more convenient way for the user who suspects the page to be
sorted by location's name.

How to reproduce:
  - Enable "Storage Locations" (inventory settings);
  - Create the following locations in this order, as children of your
    stock location:
    - Room B;
    - Room C;
    - Room A.
  - Create a new product and for this product, add 1 qty. in each of
    those newly creaed locations;
  - Create a delivery for 3 of this product, confirm and reserve;
  - Open the delivery is the Barcode App
    -> The pages' order is Room B -> Room C -> Room A.

The expected order is Room A -> Room B -> Room C.

Also, this commit moves the sorting logic outside of the creation of
lines to sort the pages when they are created (so it's explicitier).

Note about the `key` created in the `_groupLinesByPage` method and used
for the sorting:
  - The locations' name are forced in lower case because as `sort`
    compares the char's UTF-16 code to sort them, A-Z are smaller than
    a-z, so "Location D" will be placed before "Location a" if we don't
    force the string to have the same case;
  - The \x00 is the first char (the lower UTF-16 code possible).
    By using it to compare two strings, we are sure "Loc A" will always
    be placed before "Loc A/Something". Indeed, since we use the
    destination location in the string `key`, we used "_" to separate
    the source from the destination.
    But "/" is lower than "_", "Loc A/Subloc 1_Dest" is lower (and then
    placed before) than "Loc A_Dest".

task-2816768

closes odoo/enterprise#29428

X-original-commit: 8f4d9a143a874bcfe6206531b6a56a197407418a
Signed-off-by: Tiffany Chang <tic@odoo.com>