[FIX] hr_payroll_holidays: Fix reported deferred time off to next month
current behavior:
There is an action menu to automatically postpone
timeoff to the next month in the scope of the to defer mechanic.
If the time off is set on the verge of two months, it's not correctly working.

after this commit:
When you postpone a time off spitted between two months,
you only take the part on the first month to postpone it on the other month.

taskID 2810565

closes odoo/enterprise#26584

X-original-commit: 413919db59baf1bdd4b4e4e8cce4fd0b51ce6e74
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Yannick Tivisse <yti@odoo.com>