[ADD] website_sale_renting: rent on eCommerce
This commit adds rent management on eCommerce.

A rental product which is salable is always considered as rentable
unless no product pricing is defined for the website pricelist.

This module introduces few new notions:
- A rental product sold in the eCommerce cannot be picked-up nor
returned on forbidden days which are set on the company;
- A rental product sold in the eCommerce cannot be rented for a smaller
duration than the minimal duration defined on the company;

Those constraints are verified in the cart update method defined on the
sale order.

The quick add to cart button is disabled on rental products to ensure
the start and end date are always given to create the rental order line.

task-2382027