[FIX] stock_barcode: validate button tour helper
Before this commit, the helper's functions `assertValidateVisible`,
`assertValidateEnabled` and `assertValidateIsHighlighted` didn't work as
expected in some case since the OWL refactor of the `stock_barcode` app.

This commit changes those functions to work properly (and makes them
easier to write and read thanks to OWL's component system) and to work
with the "Apply" button aswell (since this button replaces the
"Validate" button in the inventory adjusment).

Also, made a minor fix in quant model' `highlightValidateButton` because
this method can highlight the apply button even when this button is
disabled.

task-2816768

X-original-commit: 5a1faa20ce4be9a1aebb344475dbf57ad1f3a20c
Part-of: odoo/enterprise#29428