[FIX] industry_fsm: don't filter records in report if admin
Before this commit, when the user is an administrator of Field Service
app, he can only see task in which he is assigned or the ones in which
he is follower of the project or those tasks. This behavior is expected
only when the user is only a basic user of this app.

This commit adds a rule for the administrator of Field Service app to
avoid filtering the records when the user is an administrator of that
app.

Related PR: #23439

X-original-commit: 5290a752f34274f8ce55b052d45c6ec919bb3a0b
Part-of: odoo/enterprise#26457