[IMP] project_enterprise,industry_fsm: compute the overlap outside SQL view
Before this commit, the report was very slowly because of the planning
overlap computed directly in the SQL view, this compute is too costly
each time we do a query in this report.

This commit removes the compute in the SQL query and made it only when
we need the result of this field. It means the field will be a compute
field non stored with a search method.

closes odoo/enterprise#26457

X-original-commit: 1c90c06a761c5480fd22cb7d3fa6a43f7f09e701
Related: odoo/odoo#89284
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>