[FIX] account_disallowed_expenses: set the translation in report account name
Step to reproduce:
	-Install accounting and the l10n_multilang and account_disallowed_expenses
	 module and add another language in the settings.
	-Create a disallowed expenses category if none exist in the accounting
	configuration
	-Select an account in chart of accounts and make a copy of an existing
	 account, change de translation of this account
	-Set a disallowed expenses category to this account
	-Create a new journal entry with this new account
	-Go in report > disallowed expenses

Current behavior:
When you click on the disallowed expenses category you have no translation and instead the name with (copy) at the end.

Expected behavior:
You should have a the translated name of the account.

For the moment the module did not take into account the possibility that the account name
was translatable (which is the case if the module l10n_multilang is installed). Hence, the
name in the report was always untranslated. This is because the report does not use the classical
ORM but uses raw SQL. This commit uses the translation of an account's name if it exists.

opw-2806289

closes odoo/enterprise#26197

X-original-commit: 17f68433eac0a8551b12ab8475bdb5fa24828028
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>