[FIX] account_sepa_direct_debit: fallback on CORE scheme if no mandate
Since commit [1], if for some reason we are trying to generate an XML file
for a payment not linked to a valid SDD mandate, a TypeError will be
raised when filling in element LclInstrm/Cd as the SDD scheme value is
inherited from the mandate associated to the payment. We should fallback
on the CORE scheme in that situation, so the real error behind is
properly raised afterward.

[1]: https://github.com/odoo/enterprise/commit/d6a6240a21457a78e49762eac5b88b84ac58f045

closes odoo/enterprise#26229

X-original-commit: 381b176bbe7eae89ab16eb7a2eedbe59489bd40f
Signed-off-by: Olivier Colson <oco@odoo.com>