[FIX] mps_mrp: Insert new content after last content
On the MPS, when updating a schedule impact another schedule not in the list, we add the impacted schedules at the end of the list.
Currently, when multiple warehouses are present in the company, the new schedule is inserted INSIDE the last schedule.
With the fix, the new schedule is inserted AFTER the last schedule.

OPW-2786188

closes odoo/enterprise#25989

X-original-commit: 9b8d0edc354b179af70860d803555dea9a6c49a0
Signed-off-by: Arnold Moyaux <arm@odoo.com>
Signed-off-by: DavidFesquet <dafr@odoo.com>