[IMP] project_account_budget: hide the budget table if no record
Before this commit, when there is no budget related to the project that
the user sees, the table and its header are displayed.

This commit removes the table when there is no record since we show no
data into that table and so this table is useless.

task-2814466

Part-of: odoo/enterprise#26004