[IMP] website_appointment: remove appointment selection when single published
PURPOSE

This commits hides the appointment type selection page when only one is
published

In many cases, having multiple appointment types is not common. When only one
is published, the appointment type selection is unnecessary. The links should
in this case redirect directly to the appointment type in question, and the
buttons used to go back to the appointment type selection should not appear.

The default query count has been incremented by one in a few tests, since
a query is added to count the number of available appointment types.

Task-2809543

closes odoo/enterprise#25777

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>