[IMP] project_account_budget: do not allow to edit budget line
Before this commit, the create and edit buttons are shown when the user
is in the form view of a budget line via the action called when the user
clicks on budgetary positions in project right side panel.

This commit hides the create and edit buttons.

task-2814466

Part-of: odoo/enterprise#26004