[FIX] l10n_cl_edi: Vendor bill acceptance - clear authentication token
after connection failure

To communicate with the SII service, Odoo automatically requests an
authentication token from it, and then uses this token in further
requests to the SII service. Sometimes, the stored authentication token
becomes invalid. When this happens, the ‘Accept Document’ and
'Claim Document' functions should request a new one. However, instead
of doing so, they simply try again with the invalid token, leading to
the error:

'Document acceptance or claim failed due to:
 - It was not possible to get a seed after 20 retries.'

This fix clears the authentcation token after a connection failure,
so that a new, valid, token will be requested when the user retries
'Accept Document' or 'Claim Document'.

closes odoo/enterprise#26830

X-original-commit: 0c93f910a2b58b986d15d36de452e65fa2541247
Signed-off-by: Josse Colpaert <jco@odoo.com>