[IMP] planning: hide 'send' and 'reset to draft' in My Planning
In the planning.slot form view in the "My Planning" menu, we hid the "Publish" and
"Publish & Send" buttons in a previous task.

However, the buttons "Reset to Draft" and "Send" for published shifts are still visible,
which is inconsistent.

This commit hides those two buttons in the "My Planning" menu.

Related: https://github.com/odoo/enterprise/pull/23423

Task-2818723

closes odoo/enterprise#26002

Signed-off-by: Xavier <xbo@odoo.com>