[FIX] project_sale_subscription: fix action used in project
Before this commit, when the user clicks on the stat button
Subscriptions in the project form view, a traceback appears because the
Query is not subscriptable, in fact, we need to convert  into a list to
get the subscription_ids fetched with the `_search` method.

This commit converts the result of the `_search` into a list to get the
ids found in the sale.subscription model.

task-2814466

Part-of: odoo/enterprise#26004