[FIX] l10n_nl_intrastat: cbs product origin country missing in export
The intrastat cbs export did not initially contain the product origin
country, as it was not required. Changes since 2022 have resulted in the
code being used for the ICV (dispatches) flow. If the end date of the
report is after 2022 and the line of the report is for a dispatch, the
product origin country code will be added to the export.

The 'country' is changed to be the intrastat country (in the case of
arrivals). This is because the country of consignment is not necessarily
the product origin country.

closes odoo/enterprise#26560

Ticket-id: 2721557
X-original-commit: 1f4c53782cfde063f77bf2c5ac8e1a21b190a6cf
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>