[FIX] industry_fsm: fix the 'planned for today' filter
Before this commit, on filter 'planned for today' only returns
the record having the start date as today.

So in this commit fixes 'planned for today' filter and returns
the record which is containing today date between the start-date and end-date.

task-2818646

X-original_commit: 614c9cb6a4febb0f8cd47457880fdfc287128b3d

closes odoo/enterprise#27585

X-original-commit: 7703935cabf4a97353b11b8ae9eecb8c00f90ec8
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>