[IMP] {website_}appointment: redesign front-end list views of appointment type
The new design includes the appointment introduction message as well as a
placeholder for the editor. The website layout is called when website_appointment
is installed.

Also, the min height of the appointment index views is set to have the website
footer just appearing. Rename id of select in the appointment list from calendarType
to appointment_type_id.

Placeholder Image appointment_svg:

Previously named Appointment.no_slot_svg, it becomes Appointment.appointment_svg.
This theme-matching helper image is now in its own file. It will be called directly
from js, replacing the rotating circle once the widget has started (if no appointments)
in both appointment selection views. The selector is renamespaced properly. The template
should be called in other views if needed. A single image prevents multiplying the
theme-matching templates. This one should be sufficient for future design.

Task-2496629

Part-of: odoo/enterprise#24744