[REM] sale_timesheet_enterprise: remove `action_billable_time_button` override
Before this commit, `action_billable_time_button` is overriden to add a
grid view and also change the form view used. But this override is no
longer useful because the action used as "template" has already the grid
view and the expected form view.

This commit just removes this override.

task-2814466

Part-of: odoo/enterprise#26004