[IMP] {website_}appointment: allow customer to bypass operator selection
Add a button 'see all availabilities' on the operator selection view,
that allows the customer to see availabilities without selecting a user.
All suggested users will be randomized and used to display the largest
amount of slots. The skip_operator_selection parameter is used to know
whether to do that or not. Chosen / selected users are also ignored in
the case of a skip. Keep queries are updated and simplified.

Task-2496629

Part-of: odoo/enterprise#24744