[IMP] project_account_budget: add 'entries' button in budget line tree
Before this commit, we do not have the `entries` in budget line tree
view when this view is not in the budget form view.

This commit adds the `entries` button in that list view to have this
button when the current clicks on a budgetary position displayed in the
budget section in project right side panel.

task-2814466

closes odoo/enterprise#26004

Related: odoo/odoo#88219
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>