[IMP] {website_}appointment: redesign the calendar time selection
The view now contains a column with the calendar view and introduction
message plus timezone selection, and one with location, duration, dynamic
selected slot date, and the slots list.

The user select is displayed only for appointment types with 'chosen'
assign_method, except if the user has chosen the 'see all availabilities'
in the operator selection. In that case we do not want the select
since we want all available users.

The title will end with 'with "name of the staff member"' and the user
select dropdown will be hidden as it makes the view cleaner: it will
happen for 'chosen' appointment types, and if a user is selected, either
with the operator selection or if there is only one. Also, we verify that
the title does not contain the username to prevent repetition.
The 'operator' chevron is clickable and appears only if the condition to
use it is fulfilled.

The 'no availability overall / this month' helpers dynamically use the
selected user's name if assign_method is 'chosen' and the operator selection
is not skipped. Redirection is only proposed when website is on.
The helper used when no availability exist in the current month is moved
inside the calendar, replacing the day names and the dates if instanciated.
It makes the design tighter and the warning appears right on the calendar
instead of much lower, (outside of the screen potentially for small devices).
On changing the tz or the user (if dropdown is there), the current month
is kept (to avoid jumping to another month as it was before this commit)

In website_appointment, we make sure the footer just slightly appears, the
same amount as other views in the appointment process. A custom editor message
on the blocks added below the calendar is added to remind they will be
available on all appointments.

The new design can be customized according to appointment properties and
selected / possible users. A single variable is used in the template of
time selection calendar: 'hide_select_dropdown' is True when the dropdown
is completely hidden, which is useful when we still want a user selection
but not to show it, for instance when a single user is available, or when
an operator has been selected in previous step.

Task-2496629

Part-of: odoo/enterprise#24744