[FIX] account_reports: hide if zero on line without balance
Steps:

Financial reports -> belgian balance sheet (bilan)
on line ASSET (ACTIF) or LIABILITY (PASSIF), check 'hide if zero'
go to belgian balance sheet
/!\ ACTIF or PASSIF are hidden
This is because we hide a line if its balance is zero, based on the
formula. In this specific case, the lines ACTIF or PASSIF have no
formulas, as their total is calculated in one of their children.

With this commit, if a line has no formulas,  we check for each line
with 'hide if zero' checked, if the sum of each column of its children
is zero, only in this case we hide the lines.

opw-2797559

closes odoo/enterprise#27852

X-original-commit: e06d22da14d506c030277d5e56885a967ab6ce82
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Guillaume Vanleynseele <guva@odoo.com>