[FIX] mass_mailing_themes: complete design-element in themes
The design tab reads styles from the `#design-element` to display the
current styles. This therefore currently requires that either all or
none of the styles are defined. In waiting for this to be remedied, this
commit therefore applies all the default styles wherever styles are
defined (in the themes).

closes odoo/enterprise#25753

Related: odoo/odoo#88318
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>