[IMP] appointment: support dash in custom URL
Currently, while sharing the appointment with custom URL, we only
support word characters (letter, digit or underscore) in the short
code. However, having dash in custom URL is nice to have and can't
hurt in any way.

With this commit, dash is supported in the custom URLs for sharing
appointments.

task-2818469

closes odoo/enterprise#26273

X-original-commit: 947259b673ab297d6fb7cbcd5648dceeb7c6443c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>