[FW][FIX] account_online_synchronization: Bank dashboard linking to wrong treeview
When clicking on the Bank dashboard hyperlink, we are redirected to
the 'Bank Statement' list while we should be on the 'Bank Statement Lines' list.

To reproduce the issue:
1. Go to Accounting/Accounting Dashboard
2. Click on the three little dots of the bank journal and then on Configuration
3. Select the "Automated Bank Synchronization" option for Bank Feeds and save
4. Go back to the Accounting Dashboard
5. Click directly on the bank journal --> the bank statement are displayed while we should see the bank statements lines.

Solution: During the refactoring (93ac6e5),
the overriden `_select_action_to_open` method in module `account_online_sync`
was removed. It should not have been.

opw-2670258

Manual Forward-Port-Of: https://github.com/odoo/enterprise/pull/25884

closes odoo/enterprise#26093

Forward-port-of: odoo/enterprise#26073
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>