[FIX] web_dashboard: update owl to last version
With the corresponding update in web, some code in web_dashboard had to
be adapted.  The main issue is that now useModel notify the views after
it is loaded, which caused the dashboard views to update its sub views
multiple times, and destroy and recreate them each time.

closes odoo/enterprise#26135

X-original-commit: 99787fcef5effb1d63c16f307a95fd0029c53539
Related: odoo/odoo#88572
Signed-off-by: Samuel Degueldre <sad@odoo.com>
Signed-off-by: Géry Debongnie <ged@odoo.com>