[FIX] website_helpdesk_livechat: ignore empty messages for ticket description
When creating a new ticket in a chat with the command '/helpdesk', the messages of
the channel are copied to the ticket description.

However, deleted messages and messages with an empty body appear in the description
as blank messages.

To avoid this, this commit adds a filter on messages that have a non empty body
when creating the ticket description.

Task-2821326

closes odoo/enterprise#26552

X-original-commit: 67e032815a1cccb199d5d8a454a3c828a8b8a9a3
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>