[FIX] account_followup: Billing access to invoice from followup
Step to reproduce:
	Install accounting
	Set a user with "Billing" rights

Current behavior:
From a followup report the user with billing rights can not
access an invoice view.

Expected behavior:
The user with billing rights has an option View Invoice when
the user is hovering the invoice in folloup report.

Solution:
We remove the groups (you must have one of this two groups if you
see this report)

opw-2811678

closes odoo/enterprise#26656

X-original-commit: 6b8cafa0b069da46b86fbe7ee48e0fd2cc19fff7
Signed-off-by: William André (wan) <wan@odoo.com>