[IMP] account_online_sync: Allows to inverse signs
In some rare cases, the provider/bank could return the wrong sign for
the balances or transactions for a specific institutions.
These new fields aim to provide a simple way to fix this issue by
allowing to inverse the sign at the time of fetching for a specific
account if needed.

Task id #2811100

closes odoo/enterprise#26117

Signed-off-by: Cedric Snauwaert <csn@odoo.com>