[IMP] account_online_sync: Do not alter confirmed statement
Sometime, we could fetch transaction that would try and impact
a confirmed bank statement. This could also happen if the provider
has an issue and provides duplicated bank statements.
Now, a new bank statement will be created if the previous one is
confirmed which would make it easy to delete it if it's an error.

Task id # 2811120

closes odoo/enterprise#26122

Signed-off-by: Cedric Snauwaert <csn@odoo.com>