[FIX] hr_payroll: fixes tracebacks in dashboard
In task 2759601 some improvements to the dashboard notes were merged. However,
after some time a change made by the js team or the editor team started causing
tracebacks in the editor with every keystroke. The main issue was that in the
initial implementation the mode of the view was changed from readonly to edit
without reinitializing the widget. This commit changes the template so that
the widget is destroyed and initialized again if the mode is changed. Also,
tests are added in order to prevent this functionality from breaking again.

task-2822711

closes odoo/enterprise#26495

X-original-commit: 162be9f1b51491345f69c2a331a42c45bb41c8a4
Signed-off-by: Kevin Baptiste <kba@odoo.com>